Move call to STS move api into project.ts (#22071)
* move call to STS move api into project.ts

* remove undefined
